src/data/requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers==4.38
matplotlib
pillow==10.2.0
huggingface-hub==0.26.5
gradio==5.9.0
numpy==1.26.4
onnxruntime
rembg
opencv-python
plyfile
safetensors
tqdm>=4.66.1
hydra-core>=1.3.2
iopath>=0.1.10